requirements.txt
fastapi
uvicorn
python-telegram-bot>=20.0
python-dotenv
google-generativeai
apscheduler
pydantic
httpx